+++ TransmissionRemote/GUI/Source/Torrent+Viewable.h
@@ -22,7 +22,7 @@ static const NSString* TorrentStatusImageNames[] = {
 
 @property (readonly) NSString *torrentIdString;
 @property (readonly) NSImage *statusImage;
-@property (readonly) NSUInteger torrentComplete;
+@property (readonly) double torrentComplete;
 @property (readonly) BOOL isDownloading;
 @property (readonly) BOOL isWaiting;
 @property (readonly) BOOL isStopping;
